packages/pulzio/backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# Auth
firebase-admin==6.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google Cloud (Secret Manager)
google-cloud-secret-manager==2.20.2

# Integrations
httpx==0.27.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.149.0

# AI
anthropic==0.39.0

# Utils
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.17
email-validator==2.2.0

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.3
